Adds an ending to the README.
Please feel free to edit this, to add your own words. An aftok
is not a faceless corporation where all communication should be
in the third person, but the sum of the efforts of all the individuals
involved, and each and every one of your voices is worthy of being
heard.